requirements.txt
aerich==0.7.2
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.29.0
attrs==23.1.0
black==23.12.0
certifi==2023.11.17
click==8.1.7
dictdiffer==0.9.0
django-environ==0.11.2
fastapi==0.105.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iso8601==1.1.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
psycopg2==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
pypika-tortoise==0.1.6
pytz==2023.3.post1
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.12.3
tortoise-orm==0.20.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
yarl==1.9.4
